Enhance data handling and metrics calculation in turbine models; update .gitignore to exclude output metrics

+++ hephaestus/single_row_models/training.py
@@ -81,7 +81,7 @@ class TabularRegressor(L.LightningModule):
             0,
             self.model.model_config.n_tokens,
             (batch_size, self.model.model_config.n_cat_cols),  # Please dont cat call
-        ).float()
+        ).long()  # Changed from .float() to .long() to match expected dtype
 
         target = torch.rand(batch_size, 1)
         return {
